separate availability checks in classes injected with DI. Checks are now uppercase named
